reverts number refactor.
(not worth the introduced complexity)

+++ barelymusician/engine/message_queue.cpp
@@ -3,14 +3,12 @@
 #include <cassert>
 #include <utility>
 
-#include "barelymusician/engine/number.h"
-
 namespace barely::internal {
 
-bool MessageQueue::Add(Integer timestamp, Message message) noexcept {
-  assert(timestamp >= 0);
-  const Integer index = write_index_;
-  const Integer next_index = (index + 1) % kMaxMessageCount;
+bool MessageQueue::Add(double timestamp, Message message) noexcept {
+  assert(timestamp >= 0.0);
+  const int index = write_index_;
+  const int next_index = (index + 1) % kMaxMessageCount;
   if (next_index == read_index_) {
     return false;
   }
@@ -19,10 +17,10 @@ bool MessageQueue::Add(Integer timestamp, Message message) noexcept {
   return true;
 }
 
-std::pair<Integer, Message>* MessageQueue::GetNext(
-    Integer end_timestamp) noexcept {
+std::pair<double, Message>* MessageQueue::GetNext(
+    double end_timestamp) noexcept {
   assert(end_timestamp >= 0.0);
-  const Integer index = read_index_;
+  const int index = read_index_;
   if (index == write_index_ || messages_[index].first >= end_timestamp) {
     return nullptr;
   }
