+++ barelymusician/engine/message_queue.h
@@ -6,7 +6,6 @@
 #include <utility>
 
 #include "barelymusician/engine/message.h"
-#include "barelymusician/engine/number.h"
 
 namespace barely::internal {
 
@@ -15,29 +14,29 @@ class MessageQueue {
  public:
   /// Adds message at timestamp.
   ///
-  /// @param timestamp Timestamp in nanoseconds.
+  /// @param timestamp Timestamp in seconds.
   /// @param message Message.
   /// @return True if successful, false otherwise.
-  bool Add(Integer timestamp, Message message) noexcept;
+  bool Add(double timestamp, Message message) noexcept;
 
   /// Returns next message before end timestamp.
   ///
-  /// @param end_timestamp End timestamp in nanoseconds.
+  /// @param end_timestamp End timestamp in seconds.
   /// @return Pointer to message if successful, nullptr otherwise.
-  std::pair<Integer, Message>* GetNext(Integer end_timestamp) noexcept;
+  std::pair<double, Message>* GetNext(double end_timestamp) noexcept;
 
  private:
   // Maximum number of messages.
-  static constexpr Integer kMaxMessageCount = 1024;
+  static constexpr int kMaxMessageCount = 1024;
 
   // List of messages with their timestamps.
-  std::array<std::pair<Integer, Message>, kMaxMessageCount> messages_;
+  std::array<std::pair<double, Message>, kMaxMessageCount> messages_;
 
   // Read index.
-  std::atomic<Integer> read_index_ = 0;
+  std::atomic<int> read_index_ = 0;
 
   // Write index.
-  std::atomic<Integer> write_index_ = 0;
+  std::atomic<int> write_index_ = 0;
 };
 
 }  // namespace barely::internal
